pyproject.toml
[project]
name = "my-researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "gpt-researcher",
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0"
]

[tool.uv.workspace]
members = [
    "backend",
]
